+++ ext/stub.h
@@ -210,6 +210,7 @@
 #include "stub/typeoff.zep.h"
 #include "stub/types/maybe.zep.h"
 #include "stub/types/mixedtype.zep.h"
+#include "stub/types/obj.zep.h"
 #include "stub/unknownclass.zep.h"
 #include "stub/unsettest.zep.h"
 #include "stub/usetest.zep.h"